chore: update hooks.server.ts for the latest @auth/sveltekit version v0.12.X. created /server/auth.ts for auth settings.
