+++ .changeset/smart-cameras-change.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+chore: update hooks.server.ts for the latest @auth/sveltekit version v0.12.X. created $lib/server/auth.ts for auth settings.
