+++ sites/geohub/package.json
@@ -36,8 +36,8 @@
 		"@neodrag/svelte": "^2.0.3",
 		"@sjmc11/tourguidejs": "^0.0.12",
 		"@sveltejs/adapter-node": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.1",
+		"@sveltejs/kit": "^2.5.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.2",
 		"@testing-library/jest-dom": "^6.2.0",
 		"@testing-library/svelte": "^4.0.5",
 		"@testing-library/user-event": "^14.5.2",
@@ -96,7 +96,7 @@
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
 		"sass": "^1.69.7",
-		"svelte": "^4.2.8",
+		"svelte": "^4.2.10",
 		"svelte-awesome-color-picker": "^2.4.8",
 		"svelte-body": "^1.4.0",
 		"svelte-check": "^3.6.2",
@@ -124,8 +124,8 @@
 		"vitest": "^1.1.1"
 	},
 	"dependencies": {
-		"@auth/core": "^0.26.0",
-		"@auth/sveltekit": "^0.11.0",
+		"@auth/core": "^0.26.3",
+		"@auth/sveltekit": "^0.12.3",
 		"@azure/service-bus": "^7.9.3",
 		"@azure/storage-blob": "^12.17.0",
 		"@azure/web-pubsub": "^1.1.1",
